services/notification-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.2.1
httpx==0.27.0
aiokafka==0.10.0
aiosmtplib==3.0.1
